style!: rename direct_aux package to directaux
The package name was failing revive linting, and also the official
package naming doc at https://go.dev/blog/package-names says not to use
underscores in package names.

Doing the package name as a standalone PR to keep the scope narrow
during a rename. There are other linter issues that need to be fixed in
a followup PR.
