+++ x/tx/signing/directaux/direct_aux_test.go
@@ -1,4 +1,4 @@
-package direct_aux_test
+package directaux_test
 
 import (
 	"context"
@@ -16,7 +16,7 @@ import (
 	signingv1beta1 "cosmossdk.io/api/cosmos/tx/signing/v1beta1"
 	txv1beta1 "cosmossdk.io/api/cosmos/tx/v1beta1"
 	"cosmossdk.io/x/tx/signing"
-	"cosmossdk.io/x/tx/signing/direct_aux"
+	"cosmossdk.io/x/tx/signing/directaux"
 )
 
 func TestDirectAuxHandler(t *testing.T) {
@@ -83,7 +83,7 @@ func TestDirectAuxHandler(t *testing.T) {
 		AuthInfoBytes: authInfoBz,
 		BodyBytes:     bodyBz,
 	}
-	modeHandler, err := direct_aux.NewSignModeHandler(direct_aux.SignModeHandlerOptions{})
+	modeHandler, err := directaux.NewSignModeHandler(directaux.SignModeHandlerOptions{})
 	require.NoError(t, err)
 
 	t.Log("verify fee payer cannot use SIGN_MODE_DIRECT_AUX")
